+++ src/test/java/com/imsweb/layout/record/csv/CommaSeparatedLayoutTest.java
@@ -12,11 +12,12 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 import org.junit.Assert;
 import org.junit.Test;
 
-import com.opencsv.CSVParserBuilder;
+import de.siegmar.fastcsv.reader.CsvReader;
 
 import com.imsweb.layout.Field.FieldAlignment;
 import com.imsweb.layout.Layout;
@@ -218,7 +219,7 @@ public class CommaSeparatedLayoutTest {
         // regular CSV fields don't have subfields; only fixed-column fields have those, so we have to tweak the design a bit...
         final Map<Integer, List<FixedColumnsField>> subFields = new HashMap<>();
         int idx = 1;
-        for (String header : new CSVParserBuilder().build().parseLine(firstLine)) {
+        for (String header : CsvReader.builder().ofCsvRecord(firstLine).stream().flatMap(f -> f.getFields().stream()).collect(Collectors.toList())) {
             FixedColumnsField field = null;
             if (header.matches("#\\d+"))
                 field = naaccrLayout.getFieldByNaaccrItemNumber(Integer.valueOf(header.substring(1)));
@@ -406,53 +407,33 @@ public class CommaSeparatedLayoutTest {
 
     @Test
     public void testReadingSpecialCases() throws IOException {
-        CommaSeparatedLayout legacyLayout = new CommaSeparatedLayout(Thread.currentThread().getContextClassLoader().getResource("testing-layout-comma-separated.xml"));
-        legacyLayout.setUseLegacyCsvParser(true);
-
         CommaSeparatedLayout rfc4180Layout = new CommaSeparatedLayout(Thread.currentThread().getContextClassLoader().getResource("testing-layout-comma-separated.xml"));
-        rfc4180Layout.setUseLegacyCsvParser(false);
 
 
-        assertCsvFields("0,123,456", legacyLayout, rfc4180Layout, "123", "456");
-        assertCsvFields("0,\"12,3\",456", legacyLayout, rfc4180Layout, "12,3", "456");
-        assertCsvFields("0,\"123\",456", legacyLayout, rfc4180Layout, "123", "456");
-        assertCsvFields("0,\"12\"\"3\"\"\",456", legacyLayout, rfc4180Layout, "12\"3\"", "456");
-        assertCsvFields("0,1\"2\"3,456", legacyLayout, rfc4180Layout, "1\"2\"3", "456");
-        assertCsvFields("0,\"1\"\"2\"\",3\",456", legacyLayout, rfc4180Layout, "1\"2\",3", "456");
+        assertCsvFields("0,123,456", rfc4180Layout, "123", "456");
+        assertCsvFields("0,\"12,3\",456", rfc4180Layout, "12,3", "456");
+        assertCsvFields("0,\"123\",456", rfc4180Layout, "123", "456");
+        assertCsvFields("0,\"12\"\"3\"\"\",456", rfc4180Layout, "12\"3\"", "456");
+        assertCsvFields("0,1\"2\"3,456", rfc4180Layout, "1\"2\"3", "456");
+        assertCsvFields("0,\"1\"\"2\"\",3\",456", rfc4180Layout, "1\"2\",3", "456");
 
-        legacyLayout.setSeparator('|');
         rfc4180Layout.setSeparator('|');
-        assertCsvFields("0|123|456", legacyLayout, rfc4180Layout, "123", "456");
-        assertCsvFields("0|\"12|3\"|456", legacyLayout, rfc4180Layout, "12|3", "456");
-        assertCsvFields("0|12,3|456", legacyLayout, rfc4180Layout, "12,3", "456");
-        assertCsvFields("0|\"123\"|456", legacyLayout, rfc4180Layout, "123", "456");
-        assertCsvFields("0|1\"2\"3|456", legacyLayout, rfc4180Layout, "1\"2\"3", "456");
-        assertCsvFields("0|\"1\"\"2\"\"|3\"|456", legacyLayout, rfc4180Layout, "1\"2\"|3", "456");
-        legacyLayout.setSeparator(',');
+        assertCsvFields("0|123|456", rfc4180Layout, "123", "456");
+        assertCsvFields("0|\"12|3\"|456", rfc4180Layout, "12|3", "456");
+        assertCsvFields("0|12,3|456", rfc4180Layout, "12,3", "456");
+        assertCsvFields("0|\"123\"|456", rfc4180Layout, "123", "456");
+        assertCsvFields("0|1\"2\"3|456", rfc4180Layout, "1\"2\"3", "456");
+        assertCsvFields("0|\"1\"\"2\"\"|3\"|456", rfc4180Layout, "1\"2\"|3", "456");
         rfc4180Layout.setSeparator(',');
 
-        // the tricky part with new lines is actually reading all the lines together for a given value; but
-        // in this test, we provide the line, and so it doesn't matter if it starts with quotes or not...
-        assertCsvFields("0,\"1\r\n2\r\n3\",456", legacyLayout, rfc4180Layout, "1\r\n2\r\n3", "456");
-        assertCsvFields("0,1\r\n2\r\n3,456", legacyLayout, rfc4180Layout, "1\r\n2\r\n3", "456");
+        //starts with quotes
+        assertCsvFields("0,\"1\r\n2\r\n3\",456", rfc4180Layout, "1\r\n2\r\n3", "456");
 
         // escape character
-        assertCsvFields("0,\"1\\2\\3\",\\", legacyLayout, rfc4180Layout, "1\\2\\3", "\\");
+        assertCsvFields("0,\"1\\2\\3\",\\", rfc4180Layout, "1\\2\\3", "\\");
         
-        // single quote, this is the only difference I could find between the legacy and RFC4180 parser (old one reports an error):
-        try {
-            legacyLayout.createRecordFromLine("0,12\"3,456", 2, null);
-            Assert.fail("Should have been an exception!");
-        }
-        catch (IOException e) {
-            Assert.assertTrue(e.getMessage().contains("Line 2"));
-        }
+        // single quote
         assertCsvFields("0,12\"3,456", rfc4180Layout, "12\"3", "456");
-    }
-
-    private void assertCsvFields(String originalLine, CommaSeparatedLayout legacyLayout, CommaSeparatedLayout rfc4180Layout, String expectedField2, String expectedField3) throws IOException {
-        assertCsvFields(originalLine, legacyLayout, expectedField2, expectedField3);
-        assertCsvFields(originalLine, rfc4180Layout, expectedField2, expectedField3);
     }
     
     private void assertCsvFields(String originalLine, CommaSeparatedLayout layout, String expectedField2, String expectedField3) throws IOException {